feat(entry admin): Improve the UX on the entry admin panel (#108)
Closes #99 
Closes #98 

I had to branch off of this to make some other changes as well
unfortunately just to get a working copy of the stack. Our expiration
was not TZ aware which broke any add (from WUI or admin). Also had to
fix the translatorConsumer because websockets were broken and spewing
errors constantly.