feat(active_blocks): Add a quick way to tell how many active blocks there are (#127)
Closes #30

This adds a pill badge to the navbar with a quick view at active blocks.
Also, it filters the homepage to only show active blocks and update the
pill badge above the actiontype list on the homepage to only show the
count of active blocks. The navbar is specific to blocks, but that seems
fair since 1) that's the main use case right now 2) that's the only
default actiontype anyways. We can always make it nicer/configurable in
the future if needed.