feat(translator): vrf support in backend (#129)
This is a bit of large MR, sorry. This MR does a few things, all of it
under the guise of supporting VRFs with GoBGP:
1) Allows you to perform all of the translator actions inside of a VRF
or on the base routing instance.
2) Split out websocket message parsing a bit and added a lot more
validation. This also means that the `process()` function is a lot more
straight-forward and that logic was broken apart.
3) We now use redis 🎉 ! This is because [GoBGP can't give you just a
single prefix's announcement
status](https://github.com/osrg/gobgp/issues/2870) so we had to write
that ourselves.
4) A ton of loggers added all over the place as well as docstrings.


closes #44, implements the translator part of #35