fix(migrations): Fix REALLY broken historicalentry migration (#138)
Wow, I screwed up. The migration default for
route_manager_historicalentry wasn't the same as the one for
route_manager_entry. Apparently, this is REALLY bad based off of how
django-simple-history works, which you can see outlined
[here](https://github.com/jazzband/django-simple-history/issues/672).

This adds a migration that nukes all history of the originating SCRAM
instance field and sets it to what the default should have been. Then
there is another migration that sets the value to what it should have
been in the first place. We also don't pull in the settings value on the
model anymore because that was also a terrible idea of mine. We now just
set that on where we create the Entry model instance.

This should fix issues related to re-blocking something that had already
been in the database. Without this fix, you'll see 500 errors when
trying to update an entry to re-block it with the following error:
`AttributeError: 'NoneType' object has no attribute
'history_change_reason'`. This is because the simple history library
tries to find the history record using ALL of the model attributes as an
exact query, so if one of the attributes on the Real Model doesn't match
what's in Historical Model, then you are in trouble land as it can't
find the model.


WHOOPSIE.

---------

Co-authored-by: Sam Oehlert <soehlert@es.net>