fix(route_manager history): move history update to after model save (#160)
We need to actually update the Entry change reason after we save the
Entry model which then creates a history model for that change. Before,
we were doing it in the serializer before the model is saved, which
means that the query to the DB that looks for the appropriate
historical_entry instance doesn't find anything because it's looking for
the historical entry that hasn't been saved yet.

This is related to #138 and actually, I think, would have fixed it, but
we came at it from a different (and frankly, I guess, wrong) angle.
Because update_change_reason only **_edits_** a historical entry, not
create a new one, but I was originally assuming that this function call
was what created the historical entry in the first place. In #138 we
"fixed" this issue by making sure that the database migrations went and
edited all the history to have the default value which then means we
technically were editing the comment for the wrong history entry.