test(restrict_changes): set this back to a 405 error
i think we do want to fail on the method, not on the user permissions. doing a 403 suggests that sometimes you can change an entry, but we always want to block PUT/PATCH as the client should just do a normal post every time and SCRAM handles the logic on our end. We don't want the client to have to track these things.