chore(deps): Set the correct postgres version everywhere and make failing checks neutral (#178)
We didn't set the postgres version in the makefile and that does seem to
matter. I'm not sure why we set it in two places, but it seems it was
done on purpose.

Also let's make the "next python version" checks neutral checks since we
know that py3.13 isn't supported until either libraries get fixed or we
upgrade to django 5.2 or newer.

Note: Until this is merged into main, checks for 3.13 on push and
pull_request will fail I think.